rename to content type classifier
