+++ nemo_curator/classifiers/__init__.py
@@ -16,10 +16,10 @@ import os
 
 os.environ["RAPIDS_NO_INITIALIZE"] = "1"
 from .aegis import AegisClassifier, InstructionDataGuardClassifier
+from .content_type import ContentTypeClassifier
 from .domain import DomainClassifier, MultilingualDomainClassifier
 from .fineweb_edu import FineWebEduClassifier
 from .quality import QualityClassifier
-from .type_of_speech import TypeOfSpeechClassifier
 
 __all__ = [
     "DomainClassifier",
@@ -28,5 +28,5 @@ __all__ = [
     "AegisClassifier",
     "InstructionDataGuardClassifier",
     "FineWebEduClassifier",
-    "TypeOfSpeechClassifier",
+    "ContentTypeClassifier",
 ]
