+++ pyproject.toml
@@ -149,7 +149,7 @@ domain_classifier_inference = "nemo_curator.scripts.classifiers.domain_classifie
 quality_classifier_inference = "nemo_curator.scripts.classifiers.quality_classifier_inference:console_script"
 aegis_classifier_inference = "nemo_curator.scripts.classifiers.aegis_classifier_inference:console_script"
 fineweb_edu_classifier_inference = "nemo_curator.scripts.classifiers.fineweb_edu_classifier_inference:console_script"
-type_of_speech_classifier_inference = "nemo_curator.scripts.classifiers.type_of_speech_classifier_inference:console_script"
+content_type_classifier_inference = "nemo_curator.scripts.classifiers.content_type_classifier_inference:console_script"
 multilingual_domain_classifier_inference = "nemo_curator.scripts.classifiers.multilingual_domain_classifier_inference:console_script"
 verify_classification_results = "nemo_curator.scripts.verify_classification_results:console_script"
 blend_datasets = "nemo_curator.scripts.blend_datasets:console_script"
